app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
ecdsa==0.19.1
fastapi==0.128.2
greenlet==3.3.1
h11==0.16.0
idna==3.11
nest-asyncio==1.6.0
numpy==2.4.1
packaging==25.0
parso==0.8.5
passlib==1.7.4
platformdirs==4.5.1
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.2.14